requirements-dev.txt
Faker==8.13.2
GitPython==3.1.17
boto3==1.18.45
pylint==2.10.2
absl-py==0.13.0
protobuf==3.17.3
numpy==1.21.2
pyarrow==5
pandas==1.3.3
clickhouse-sqlalchemy==0.1.6
antlr4-python3-runtime==4.9.3
graphviz==0.19.1
